requirements.txt
requests==2.32.2
Pillow==10.4.0
font-hanken-grotesk==0.0.2
font-intuitive==0.0.4
inky==1.5.0; platform_machine == "armv7l" or platform_machine == "aarch64"
numpy==2.0.0
pytest
pyyaml